done with the add record amount

+++ internal/contract.go
@@ -20,7 +20,7 @@ type (
 		UserGUID    uuid.UUID `json:"user_guid" db:"user_guid"`
 		Category    string    `json:"category" db:"category"`
 		Description string    `json:"description" db:"description"`
-		Amount      float64   `json:"amount" db:"amount"`
+		Amount      uint32    `json:"amount" db:"amount"`
 		CreatedAt   time.Time `json:"created_at" db:"created_at"`
 		UpdatedAt   time.Time `json:"updated_at" db:"updated_at"`
 	}
@@ -28,7 +28,7 @@ type (
 	SpendingRecord struct {
 		GUID         uuid.UUID `json:"guid" db:"guid"`
 		CategoryGUID uuid.UUID `json:"category_guid" db:"category_guid"`
-		Amount       float32   `json:"amount" db:"amount"`
+		Amount       uint16    `json:"amount" db:"amount"`
 		Description  string    `json:"description" db:"description"`
 		CreatedAt    time.Time `json:"created_at" db:"created_at"`
 		UpdatedAt    time.Time `json:"updated_at" db:"updated_at"`
